backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
insightface==0.7.3
faiss-cpu==1.7.4
torch==2.1.0
torchvision==0.16.0
python-dotenv==1.0.0
aiofiles==0.24.0
websockets==12.0
email-validator==2.1.0
Pillow==10.1.0
psutil==5.9.6